Группировка rest api, удаление из корневого /cards/ большого объема данных и вынос в другую цель
