+++ backend/restapi/app/cards/rules/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from restapi.app.cards.rules.models import CardRule
+
+class CardRuleSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = CardRule
+        fields = ["id", "date", "text"]
